requirements.txt
fastapi~=0.115
loguru~=0.7
pathlib~=1.0
setuptools~=75.8
uvicorn~=0.34
pydantic-settings~=2.9
PyYAML>=6.0
python-dotenv
psycopg2-binary
sqlalchemy